Move body of ScriptTask::handle_mouse_move_event into a method on Document #4984
